The information from the JSON files returned by the PokeApi have been separated into separate data classes that will allow specific Pokemon attributes to be added. This is also a good forking point as from here I will narrow down the Pokemon attributes actually show in the app, so if I wish to come back and change which attributes I display, this would be a good return point.
Created a retrofit api interface that defines the functions for network requests.

Setup the Pokemon repository with basic error handling (this could be done in more detail later).

Create a constants object that contains the base api url so if for whatever reason the url changes in the future this can easily be handled by amending one line of code.

DaggerHilt provider functions were set up in the AppModule to allow for future dependency injection into Kotlin classes (viewmodels).